requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
gspread>=6.0.0
google-auth>=2.30.0
google-cloud-storage>=2.18.0
google-cloud-documentai>=2.31.0
twilio>=9.0.0
reportlab>=4.2.0
